restructure files and added sass color variables
