+++ components/Nav.js
@@ -1,0 +1,19 @@
+import Link from 'next/link'
+// import navStyles from '../styles/Nav.module.css'
+
+const Nav = () => {
+  return (
+    <nav className="container">
+      <ul>
+        <li>
+          <Link href="/">Home</Link>
+        </li>
+        <li>
+          <Link href="/about">About</Link>
+        </li>
+      </ul>
+    </nav>
+  )
+}
+
+export default Nav
