+++ pages/_app.tsx
@@ -1,10 +1,12 @@
-import { AppProps } from 'next/app'
-import '../styles/globals.scss'
+import "../styles/globals.scss";
+import Layout from "../components/Layout";
 
-
-function MyApp({ Component, pageProps }: AppProps) {
-
-  return <Component {...pageProps} />
+function MyApp({ Component, pageProps }) {
+  return (
+    <Layout>
+      <Component {...pageProps} />
+    </Layout>
+  );
 }
 
-export default MyApp
+export default MyApp;
