--- styles/Home.module.scss
@@ -1,10 +1,0 @@
-@value colors: './colors.scss';
-@value primary-base, primary-light, primary-dark, secondary-base, secondary-light, tertiary-base, fourth-base, fifth-base, highlight, neutral, background from colors;
-
-.profilePic {
-  max-height: 150px;
-  max-width: 150px;
-  border-radius: 100%;
-  margin-bottom: 20px;
-  border: 3px solid primary-base;
-}
