+++ styles/SpotlightCard.module.scss
@@ -5,7 +5,6 @@
   margin-bottom: 20px;
 }
 
-
 .icon {
   height: 25px;
   width: 25px;
