+++ styles/globals.scss
@@ -1,5 +1,7 @@
-@import "~bootstrap/scss/bootstrap.scss";
+@import '../node_modules/bootstrap/scss/bootstrap';
+@import './colors.scss';
 
+//global styles
 html,
 body {
   padding: 0;
@@ -16,7 +18,6 @@ a {
 * {
   box-sizing: border-box;
 }
-
 
 .icon-small {
   width: 50px;
